Rename ttnt_module.rb to internals.rb
