+++ lib/ttnt/internals.rb
